Rough implementation of Character class

+++ Character_header/Character.h
@@ -20,11 +20,11 @@ public:
 Character::Character();
 int getHealth () const;
 CharType getType();
-void setHealth(double);
+void setHealth(int);
 bool isalive() const ;
 Character(CharType, const string &, int, int, int);
-void damage(double dam);
-void takeDamage(double);
+void damage(int dam);
+void takeDamage(int);
 void attack(Character &);
 
 };
